🚑️ Support lowercase and UPPERCASE for Enums
The CodeTF JSON schema says that enums can use lowercase or uppercase.
This change makes the enums prefer lowercase (i.e. in serialization),
but they will accept uppercase (i.e. in deserialization). This is a good
transition towards removing support for UPPERCASE if we would like to
take that next step.

/towards ISS-617